refactor(storage): relocate EventFilter/FactScoringRow/SessionFact to types
P4 of #629. These are dialect-free domain rows (a Postgres backend returns
them too), so their structural home is `types` (the dependency leaf), not the
SQLite `store` module — the storage port must not reference a type living
inside the impl it abstracts. `pub use` shims in store::{events,facts} preserve
the original paths (e.g. forgetting::policy keeps its FactScoringRow import +
ImportanceInputs impl). Behavior-neutral: the full suite stays green.

Part of #629.
